Day 04 validation with Python
I wanted to try out how it feels to treat a 2D grid as a complex plane.
It makes it trivial to move in any direction as we are just adding
complex numbers instead of having to deal with separate coordinates!
Unfortunately acton does not natively support complex numbers yet so the
first implementation of this is in Python :o